include plan refs in item use count
