+++ src/main/java/com/brennaswitzer/cookbook/repositories/PantryItemSearchRepository.java
@@ -5,6 +5,8 @@ import com.brennaswitzer.cookbook.repositories.impl.PantryItemSearchRequest;
 
 public interface PantryItemSearchRepository {
 
+    long countTotalUses(PantryItem pantryItem);
+
     SearchResponse<PantryItem> search(PantryItemSearchRequest request);
 
 }
