+++ src/main/java/com/brennaswitzer/cookbook/repositories/RecipeSearchRepositoryImpl.java
@@ -1,8 +1,6 @@
 package com.brennaswitzer.cookbook.repositories;
 
-import com.brennaswitzer.cookbook.domain.PantryItem;
 import com.brennaswitzer.cookbook.domain.Recipe;
-import com.brennaswitzer.cookbook.domain.User;
 import com.brennaswitzer.cookbook.repositories.impl.LibrarySearchRequest;
 import com.brennaswitzer.cookbook.util.NamedParameterQuery;
 import com.brennaswitzer.cookbook.util.UserPrincipalAccess;
@@ -104,37 +102,6 @@ public class RecipeSearchRepositoryImpl implements RecipeSearchRepository {
         List<Recipe> resultList = (List<Recipe>) query.getResultList();
 
         return SearchResponse.of(request, resultList);
-    }
-
-    @Override
-    public long countTotalUses(PantryItem pantryItem) {
-        if (pantryItem.getUseCount() != null) {
-            return pantryItem.getUseCount();
-        }
-        return countUses(null, pantryItem);
-    }
-
-    @Override
-    public long countMyUses(PantryItem pantryItem) {
-        return countUses(principalAccess.getUser(), pantryItem);
-    }
-
-    private long countUses(User owner, PantryItem pantryItem) {
-        var q = new NamedParameterQuery(
-                """
-                select count(distinct r.id)
-                from Recipe r
-                    join r.ingredients ing
-                where ing.ingredient.id = :id
-                """, "id", pantryItem.getId());
-        if (owner != null) {
-            q.append("  and r.owner.id = :owner",
-                     "owner",
-                     owner.getId());
-        }
-        var query = entityManager.createQuery(q.getStatement());
-        q.forEachParameter(query::setParameter);
-        return (Long) query.getResultList().iterator().next();
     }
 
 }
